+++ src/main/java/kitchenpos/deliveryorders/domain/KitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.orders.deliveryorders.domain;
+package kitchenpos.deliveryorders.domain;
 
 import java.math.BigDecimal;
 import java.util.UUID;
