+++ src/main/java/kitchenpos/eatinorders/application/OrderTableService.java
@@ -1,9 +1,9 @@
-package kitchenpos.orders.eatinorders.application;
+package kitchenpos.eatinorders.application;
 
-import kitchenpos.orders.legacy.domain.OrderRepository;
-import kitchenpos.orders.legacy.domain.OrderStatus;
-import kitchenpos.orders.eatinorders.domain.order_table.OrderTable;
-import kitchenpos.orders.eatinorders.domain.order_table.OrderTableRepository;
+import kitchenpos.eatinorders.domain.order_table.OrderTable;
+import kitchenpos.legacy_orders.domain.OrderRepository;
+import kitchenpos.legacy_orders.domain.OrderStatus;
+import kitchenpos.eatinorders.domain.order_table.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
