+++ src/main/java/kitchenpos/eatinorders/infra/JpaOrderTableRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.orders.eatinorders.infra;
+package kitchenpos.eatinorders.infra;
 
-import kitchenpos.orders.eatinorders.domain.order_table.OrderTable;
-import kitchenpos.orders.eatinorders.domain.order_table.OrderTableRepository;
+import kitchenpos.eatinorders.domain.order_table.OrderTable;
+import kitchenpos.eatinorders.domain.order_table.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
