+++ src/main/java/kitchenpos/legacy_orders/domain/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.orders.legacy.domain;
+package kitchenpos.legacy_orders.domain;
 
 public enum OrderStatus {
     WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
