+++ src/main/java/kitchenpos/legacy_orders/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.orders.legacy.domain;
+package kitchenpos.legacy_orders.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
