+++ src/test/java/kitchenpos/application/InMemoryOrderRepository.java
@@ -1,9 +1,9 @@
 package kitchenpos.application;
 
-import kitchenpos.orders.legacy.domain.Order;
-import kitchenpos.orders.legacy.domain.OrderRepository;
-import kitchenpos.orders.legacy.domain.OrderStatus;
-import kitchenpos.orders.eatinorders.domain.order_table.OrderTable;
+import kitchenpos.legacy_orders.domain.Order;
+import kitchenpos.legacy_orders.domain.OrderRepository;
+import kitchenpos.legacy_orders.domain.OrderStatus;
+import kitchenpos.eatinorders.domain.order_table.OrderTable;
 
 import java.util.ArrayList;
 import java.util.HashMap;
