+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,14 +1,14 @@
 package kitchenpos.application;
 
 import kitchenpos.menu.domain.menu.MenuRepository;
-import kitchenpos.orders.legacy.application.OrderService;
-import kitchenpos.orders.legacy.domain.Order;
-import kitchenpos.orders.legacy.domain.OrderLineItem;
-import kitchenpos.orders.legacy.domain.OrderRepository;
-import kitchenpos.orders.legacy.domain.OrderStatus;
-import kitchenpos.orders.eatinorders.domain.order_table.OrderTable;
-import kitchenpos.orders.eatinorders.domain.order_table.OrderTableRepository;
-import kitchenpos.orders.legacy.domain.OrderType;
+import kitchenpos.legacy_orders.application.OrderService;
+import kitchenpos.legacy_orders.domain.Order;
+import kitchenpos.legacy_orders.domain.OrderLineItem;
+import kitchenpos.legacy_orders.domain.OrderRepository;
+import kitchenpos.legacy_orders.domain.OrderStatus;
+import kitchenpos.eatinorders.domain.order_table.OrderTable;
+import kitchenpos.eatinorders.domain.order_table.OrderTableRepository;
+import kitchenpos.legacy_orders.domain.OrderType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
