+++ src/test/java/kitchenpos/application/OrderTableServiceTest.java
@@ -1,10 +1,10 @@
 package kitchenpos.application;
 
-import kitchenpos.orders.legacy.domain.OrderRepository;
-import kitchenpos.orders.legacy.domain.OrderStatus;
-import kitchenpos.orders.eatinorders.application.OrderTableService;
-import kitchenpos.orders.eatinorders.domain.order_table.OrderTable;
-import kitchenpos.orders.eatinorders.domain.order_table.OrderTableRepository;
+import kitchenpos.legacy_orders.domain.OrderRepository;
+import kitchenpos.legacy_orders.domain.OrderStatus;
+import kitchenpos.eatinorders.application.OrderTableService;
+import kitchenpos.eatinorders.domain.order_table.OrderTable;
+import kitchenpos.eatinorders.domain.order_table.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
